chore: remove useless split source function
