+++ packages/fs-router-vite-plugin/src/constants.ts
@@ -1,1 +1,3 @@
-export const SPLIT_PREFIX = 'sp-'
+export const ROUTES_FOLDER = './src/routes/'
+export const ROOT_PATH_ID = '__root'
+export const GENERATED_ROUTE_TREE = './.tuono/routeTree.gen.ts'
