+++ packages/fs-router-vite-plugin/src/index.ts
@@ -1,10 +1,6 @@
-import { fileURLToPath, pathToFileURL } from 'url'
-import { routeGenerator } from './routes-generator'
+import { routeGenerator } from './generator'
 
 import { normalize } from 'path'
-// eslint-disable-next-line sort-imports
-import { makeCompile, splitFile } from './compiler'
-import { SPLIT_PREFIX } from './constants'
 
 import type { Plugin } from 'vite'
 
@@ -12,15 +8,15 @@ const ROUTES_DIRECTORY_PATH = './src/routes'
 
 let lock = false
 
-export function RouterGenerator(): Plugin {
+export default function RouterGenerator(): Plugin {
   const generate = async (): Promise<void> => {
     if (lock) return
     lock = true
 
     try {
-      // TODO: generator function
       await routeGenerator()
     } catch (err) {
+      console.error(err)
     } finally {
       lock = false
     }
@@ -30,13 +26,12 @@ export function RouterGenerator(): Plugin {
     const filePath = normalize(file)
 
     if (filePath.startsWith(ROUTES_DIRECTORY_PATH)) {
-      // TODO: generator function
       await generate()
     }
   }
 
   return {
-    name: 'vite-plugin-tuono-fs-router-generator',
+    name: 'vite-plugin-tuono-fs-router',
     configResolved: async (): Promise<void> => {
       await generate()
     },
@@ -49,42 +44,4 @@ export function RouterGenerator(): Plugin {
       }
     },
   }
-}
-
-export function RouterCodeSplitter(): Plugin {
-  const ROOT: string = process.cwd()
-
-  return {
-    name: 'vite-plugin-tuono-fs-router-code-splitter',
-    enforce: 'pre',
-    resolveId(source): string | null {
-      if (source.startsWith(SPLIT_PREFIX + ':')) {
-        return source.replace(SPLIT_PREFIX + ':', '')
-      }
-      return null
-    },
-    async transform(code, id): Promise<any> {
-      const url = pathToFileURL(id)
-      url.searchParams.delete('v')
-      id = fileURLToPath(url).replace(/\\/g, '/')
-
-      const compile = makeCompile({ root: ROOT })
-
-      if (id.includes(SPLIT_PREFIX)) {
-        const compiled = await splitFile({
-          code,
-          compile,
-          filename: id,
-        })
-
-        return compiled
-      }
-
-      return null
-    },
-  }
-}
-
-export function ViteFsRouter(): Plugin[] {
-  return [RouterGenerator(), RouterCodeSplitter()]
 }
