+++ packages/fs-router-vite-plugin/src/types.ts
