+++ packages/fs-router-vite-plugin/src/utils.ts
